backend/requirements.txt
#Framework para la API
fastapi
uvicorn[standard]
aiohttp

#Orquestación del Agente
langchain
langgraph
langchain-community
langchain-openai  # O el proveedor de LLM que prefieras (e.g., langchain-azure-openai)

#Conectores de Azure y Databricks
azure-cosmos
azure-search-documents
databricks-sql-connector
azure-storage-blob==12.26.0
python-dotenv
pyarrow==21.0.0

#Modelado de datos
pydantic